nano-e-commerce/services/api-gateway/requirements.txt
fastapi==0.105.0
uvicorn==0.25.0
grpcio==1.62.0
grpcio-tools==1.62.0
protobuf==4.25.2
pydantic==2.6.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
asyncpg==0.29.0
redis==5.0.1
